[AIRFLOW-4536] Split jobs into package
